requirements.txt
asgiref==3.8.1
catboost==1.2.8
contourpy==1.3.2
cycler==0.12.1
Django==5.2.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Faker==37.4.0
fonttools==4.58.4
graphviz==0.21
joblib==1.5.1
kiwisolver==1.4.8
lightgbm==4.6.0
matplotlib==3.10.3
narwhals==1.43.1
numpy==2.3.0
packaging==25.0
pandas==2.3.0
pillow==11.2.1
plotly==6.1.2
psycopg2-binary==2.9.10
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-learn==1.7.0
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tzdata==2025.2
